refactor: MessageRoomException 으로 통일
#524
