+++ backend/emm-sale/src/main/java/com/emmsale/message_room/exception/MessageRoomExceptionType.java
@@ -9,8 +9,8 @@ public enum MessageRoomExceptionType implements BaseExceptionType {
       HttpStatus.NOT_FOUND,
       "해당 쪽지방은 존재하지 않습니다."
   ),
-
-  ;
+  SENDER_IS_NOT_EQUAL_REQUEST_MEMBER(HttpStatus.FORBIDDEN,
+      "로그인한 사용자와 메세지를 보내는 사용자가 다릅니다.");
 
   private final HttpStatus httpStatus;
   private final String errorMessage;
